Refactor, rename, cleanup
